scaffold separate stdlib and wasm features
