+++ Cargo.toml
@@ -11,6 +11,12 @@ description = "The simplest CMS in existence"
 name = "archival"
 path = "src/lib.rs"
 
+[features]
+default = ["stdlib-fs", "binary"]
+binary = ["dep:notify", "dep:ctrlc"]
+stdlib-fs = ["dep:fs_extra", "dep:walkdir"]
+wasm-fs = ["dep:web-sys"]
+
 [[bin]]
 name = "archival"
 path = "src/main.rs"
@@ -19,14 +25,26 @@ path = "src/main.rs"
 liquid = "0.26.1"
 serde = "1.0.152"
 toml = "0.7.1"
-walkdir = "2.3.3"
 comrak = "0.18"
 liquid-core = "0.26.4"
 liquid-lib = { version = "0.26.4", features = ["shopify", "extra"] }
 regex = "1.10.2"
-fs_extra = "1.3.0"
-notify = "6.1.1"
-ctrlc = "3.4.2"
+# stdlib-fs
+walkdir = {version = "2.3.3", optional = true}
+fs_extra = {version = "1.3.0", optional = true}
+notify = {version = "6.1.1", optional = true}
+ctrlc = {version = "3.4.2", optional = true}
+# wasm-fs
+js-sys = {version = "0.3.66", optional = true}
+web-sys = {version = "0.3.66", optional = true, features = [
+    "FileSystem",
+    "FileSystemEntry",
+    "FileSystemFlags",
+    "FileSystemDirectoryEntry",
+    "FileSystemDirectoryEntry",
+    "FileSystemEntryCallback",
+]}
 
 [dev-dependencies]
 tempfile = "3.5.0"
+wasm-bindgen-test = "0.3.0"
